// src/models/plant.d.ts
export declare const PlantSchema: z.ZodObject<{
    id: z.ZodString;
    name: z.ZodString;
    species: z.ZodString;
    location: z.ZodOptional<z.ZodString>;
    plantedDate: z.ZodOptional<z.ZodString>;
    lastWatered: z.ZodOptional<z.ZodString>;
    wateringFrequency: z.ZodOptional<z.ZodNumber>;
    lastFertilized: z.ZodOptional<z.ZodString>;
    fertilizingFrequency: z.ZodOptional<z.ZodNumber>;
    notes: z.ZodOptional<z.ZodString>;
    sunlightRequirement: z.ZodOptional<z.ZodEnum<["full-sun", "partial-shade", "shade"]>>;
    soilType: z.ZodOptional<z.ZodString>;
    harvestDate: z.ZodOptional<z.ZodString>;
    createdAt: z.ZodString;
    updatedAt: z.ZodString;
}, "strip", z.ZodTypeAny, {
    id: string;
    name: string;
    species: string;
    createdAt: string;
    updatedAt: string;
    location?: string | undefined;
    plantedDate?: string | undefined;
    lastWatered?: string | undefined;
    wateringFrequency?: number | undefined;
    lastFertilized?: string | undefined;
    fertilizingFrequency?: number | undefined;
    notes?: string | undefined;
    sunlightRequirement?: "full-sun" | "partial-shade" | "shade" | undefined;
    soilType?: string | undefined;
    harvestDate?: string | undefined;
}, {
    id: string;
    name: string;
    species: string;
    createdAt: string;
    updatedAt: string;
    location?: string | undefined;
    plantedDate?: string | undefined;
    lastWatered?: string | undefined;
    wateringFrequency?: number | undefined;
    lastFertilized?: string | undefined;
    fertilizingFrequency?: number | undefined;
    notes?: string | undefined;
    sunlightRequirement?: "full-sun" | "partial-shade" | "shade" | undefined;
    soilType?: string | undefined;
    harvestDate?: string | undefined;
}>;
export type Plant = z.infer<typeof PlantSchema>;